multiple bracket handle done

+++ src/soumyagh/expeval/cli/ExpEvalMain.java
@@ -5,7 +5,7 @@ import soumyagh.expeval.lib.ExpEvaluator;
 public class ExpEvalMain {
     public static void main(String[] args) throws Exception {
         ExpEvaluator eval = new ExpEvaluator();
-        int result = eval.getResultByEvaluating(args);
+        double result = eval.getResultByEvaluating(args);
         System.out.println(result);
     }
 }
